+++ shared/chat/conversation/messages/message-popup/exploding/container.tsx
@@ -123,7 +123,7 @@ export default Container.connect(
     _onKick: (teamID: TeamTypes.TeamID, username: string) =>
       dispatch(
         RouteTreeGen.createNavigateAppend({
-          path: [{props: {navToChat: true, teamID, username}, selected: 'teamReallyRemoveMember'}],
+          path: [{props: {members: [username], teamID}, selected: 'teamReallyRemoveMember'}],
         })
       ),
     _onPinMessage: () => {
