+++ shared/chat/conversation/messages/message-popup/text/container.tsx
@@ -100,7 +100,7 @@ const mapDispatchToProps = (dispatch: Container.TypedDispatch) => ({
   _onKick: (teamID: TeamTypes.TeamID, username: string) =>
     dispatch(
       RouteTreeGen.createNavigateAppend({
-        path: [{props: {navToChat: true, teamID, username}, selected: 'teamReallyRemoveMember'}],
+        path: [{props: {members: [username], teamID}, selected: 'teamReallyRemoveMember'}],
       })
     ),
   _onPinMessage: (message: Types.Message) => {
