+++ shared/teams/stories/index.tsx
@@ -5,6 +5,8 @@ import * as Constants from '../../constants/teams'
 export const fakeTeamID = 'fakeTeamID'
 const teamID2 = 'faketeamID2'
 const teamID3 = 'faketeamID3'
+const subteam1 = 'subteam1'
+const subteam2 = 'subteam2'
 export const fakeTeamIDs = [fakeTeamID, teamID2, teamID3]
 
 export const store = Container.produce(Sb.createStoreWithCommon(), draftState => {
@@ -22,6 +24,7 @@ export const store = Container.produce(Sb.createStoreWithCommon(), draftState =>
             ['paula', {fullName: 'Paula Superlonglastnamelikereallylongforreal', status: 'active', type: 'writer', username: 'paula'}],
             ['andonuts', {fullName: '', status: 'active', type: 'writer', username: 'andonuts'}],
           ]),
+          subteams: new Set([subteam1, subteam2]),
         },
       ],
       [
@@ -38,6 +41,11 @@ export const store = Container.produce(Sb.createStoreWithCommon(), draftState =>
       [fakeTeamID, Constants.makeTeamMeta({memberCount: 32, teamname: 'keybase_storybook'})],
       [teamID2, Constants.makeTeamMeta({isOpen: true, memberCount: 11947, teamname: 'fan_club'})],
       [teamID3, Constants.makeTeamMeta({isOpen: false, memberCount: 234, teamname: 'club_penguin'})],
+      [
+        subteam1,
+        Constants.makeTeamMeta({isOpen: true, memberCount: 980, teamname: 'keybase_storybook.friends'}),
+      ],
+      [subteam2, Constants.makeTeamMeta({isOpen: false, memberCount: 7, teamname: 'keybase_storybook.exec'})],
     ]),
     teamRoleMap: {
       ...draftState.teams.teamRoleMap,
